cambio de float a int

+++ carrito/context_processor.py
@@ -1,6 +1,6 @@
 def monto_total_cart(request):
-    total = 0.0
+    total = 0
     if request.user.is_authenticated:
         for key, value in request.session['cart'].items():
-            total = total + (float(value['precio']) * value['cantidad'])
+            total = total + (int(value['precio']) * value['cantidad'])
     return {'monto_total_cart': total}
