+++ carrito/templatetags/cart_tags.py
@@ -4,7 +4,7 @@ register = template.Library()
 
 @register.filter()
 def multiplicar(value, arg):
-    return float(value) * arg
+    return int(value) * arg
 
 @register.filter()
 def formato_monetario(value, arg):
